uint64 fix

+++ modules/workers/FSWPollWorker.js
@@ -507,7 +507,7 @@ function poll(aArgs) {
 				var fsstream;
 				
 				if (!('cfArrRef' in macStuff)) {
-					macStuff.cStr_ptrOf_cfArrRef = ctypes.char.array(macStuff.maxLenCfArrRefPtrStr).ptr(aArgs.ptrStrOf__cStr_ptrOf_cfArrRef);
+					macStuff.cStr_ptrOf_cfArrRef = ctypes.char.array(macStuff.maxLenCfArrRefPtrStr).ptr(ctypes.UInt64(aArgs.ptrStrOf__cStr_ptrOf_cfArrRef));
 				}
 				if (!('cId' in macStuff)) {
 					macStuff.cId = ostypes.API('FSEventsGetCurrentEventId')(); // ostypes.TYPE.FSEventStreamEventId(ostypes.CONST.kFSEventStreamEventIdSinceNow);
